AppSettings: fix crash when trying to delete database, but directory with path MediaSelectorElement.MediaPath not existing

+++ DLR_Data_App/DLR_Data_App/DLR_Data_App/Views/Settings/AppSettings.xaml.cs
@@ -44,7 +44,9 @@ namespace DLR_Data_App.Views.Settings
                     answer = await DisplayAlert(AppResources.removedatabase, AppResources.removedatabasewarning, AppResources.accept, AppResources.cancel);
                     if(answer)
                     {
-                        Directory.Delete(MediaSelectorElement.MediaPath, true);
+                        if (Directory.Exists(MediaSelectorElement.MediaPath))
+                            Directory.Delete(MediaSelectorElement.MediaPath, true);
+
                         if(Database.RemoveDatabase())
                         {
                             // Successful
